+++ src/ywrenumberlib/config_renumber.py
@@ -26,8 +26,8 @@ class ConfigRenumber:
         Required keyword arguments:
             ren_unused -- bool: include chapters marked "Unused" in yWriter.
             ren_parts -- bool: include chapters marked "This chapter begins a new section" in yWriter.
-            consider_novel_start -- bool: start numbering with the chapter marked as "start of novel" in yWriter.
-            renumber_within_parts -- bool: Reset the chapter number after section beginnings.
+            ren_regular -- bool: start numbering with the chapter marked as "start of novel" in yWriter.
+            ren_within_parts -- bool: Reset the chapter number after section beginnings.
             numbering_style -- str: '0'=Arabic numbers; '1'= Roman numbers; '2'= Written out in English.
             numbering_case -- str: '0'=Uppercase; '1'=Capitalized; '2'=Lowercase.
             heading_prefix -- str: a string preceding each number.
@@ -37,6 +37,11 @@ class ConfigRenumber:
         row1Cnt = 1
         hdTypes = tk.Label(window, text='Chapters')
         hdTypes.grid(row=row1Cnt, column=1, sticky=tk.W, padx=20)
+        row1Cnt += 1
+        self._renRegular = tk.BooleanVar(value=kwargs['ren_regular'])
+        startCheckbox = ttk.Checkbutton(window, text='Include regular chapters',
+                                         variable=self._renRegular, onvalue=True, offvalue=False)
+        startCheckbox.grid(row=row1Cnt, column=1, sticky=tk.W, padx=20)
         row1Cnt += 1
         self._renParts = tk.BooleanVar(value=kwargs['ren_parts'])
         partsCheckbox = ttk.Checkbutton(window, text='Include section beginnings',
@@ -48,12 +53,7 @@ class ConfigRenumber:
                                          variable=self._renUnused, onvalue=True, offvalue=False)
         unusedCheckbox.grid(row=row1Cnt, column=1, sticky=tk.W, padx=20)
         row1Cnt += 1
-        self._considerStart = tk.BooleanVar(value=kwargs['consider_novel_start'])
-        startCheckbox = ttk.Checkbutton(window, text='Begin at "start of novel"',
-                                         variable=self._considerStart, onvalue=True, offvalue=False)
-        startCheckbox.grid(row=row1Cnt, column=1, sticky=tk.W, padx=20)
-        row1Cnt += 1
-        self._partWise = tk.BooleanVar(value=kwargs['renumber_within_parts'])
+        self._partWise = tk.BooleanVar(value=kwargs['ren_within_parts'])
         partWiseCheckbox = ttk.Checkbutton(window, text='Reset number after section beginnings',
                                          variable=self._partWise, onvalue=True, offvalue=False)
         partWiseCheckbox.grid(row=row1Cnt, column=1, sticky=tk.W, padx=20)
@@ -109,10 +109,10 @@ class ConfigRenumber:
         Positional arguments:
             kwargs -- reference to the ui kwargs dictionary.
         """
-        kwargs['ren_parts'] = self._renParts.get()
+        kwargs['ren_regular'] = self._renRegular.get()
         kwargs['ren_unused'] = self._renUnused.get()
-        kwargs['consider_novel_start'] = self._considerStart.get()
-        kwargs['renumber_within_parts'] = self._partWise.get()
+        kwargs['ren_parts'] = self._renParts.get()
+        kwargs['ren_within_parts'] = self._partWise.get()
         kwargs['numbering_style'] = str(self._numberingStyle.get())
         kwargs['numbering_case'] = str(self._numberingCase.get())
         kwargs['heading_prefix'] = f'|{self._headingPrefix.get()}|'